feat : split testcontainers config for reusability
