+++ src/test/java/com/learning/mfscreener/repository/SchemaValidationPostgresTest.java
@@ -2,24 +2,23 @@ package com.learning.mfscreener.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import jakarta.persistence.EntityManager;
+import com.learning.mfscreener.common.SQLContainersConfig;
+import com.zaxxer.hikari.HikariDataSource;
+import javax.sql.DataSource;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.annotation.Import;
 
-@DataJpaTest(
-        properties = {
-            "spring.jpa.hibernate.ddl-auto=validate",
-            "spring.test.database.replace=none",
-            "spring.datasource.url=jdbc:tc:postgresql:16-alpine:///db"
-        })
+@DataJpaTest(properties = {"spring.jpa.hibernate.ddl-auto=validate", "spring.test.database.replace=none"})
+@Import(SQLContainersConfig.class)
 class SchemaValidationPostgresTest {
 
     @Autowired
-    EntityManager entityManager;
+    DataSource dataSource;
 
     @Test
-    void schemaValidity() {
-        assertThat(entityManager).isNotNull();
+    void validateJpaMappingsWithDbSchema() {
+        assertThat(dataSource).isNotNull().isInstanceOf(HikariDataSource.class);
     }
 }
